relax assert after new highway release
