+++ lib/jxl/fast_math_test.cc
@@ -107,6 +107,22 @@ HWY_NOINLINE void TestFastErf() {
   printf("max abs err %e\n", static_cast<double>(max_abs_err));
 }
 
+HWY_NOINLINE void TestCubeRoot() {
+  const HWY_FULL(float) d;
+  for (uint64_t x5 = 0; x5 < 2000000; x5++) {
+    const float x = x5 * 1E-5f;
+    const float expected = cbrtf(x);
+    HWY_ALIGN float approx[MaxLanes(d)];
+    Store(CubeRootAndAdd(Set(d, x), Zero(d)), d, approx);
+
+    // All lanes are same
+    for (size_t i = 1; i < Lanes(d); ++i) {
+      EXPECT_NEAR(approx[0], approx[i], 0);
+    }
+    EXPECT_NEAR(approx[0], expected, 8E-7f);
+  }
+}
+
 HWY_NOINLINE void TestFastSRGB() {
   constexpr size_t kNumTrials = 1 << 23;
   Rng rng(1);
@@ -261,6 +277,7 @@ HWY_EXPORT_AND_TEST_P(FastMathTargetTest, TestFastPow2);
 HWY_EXPORT_AND_TEST_P(FastMathTargetTest, TestFastPow);
 HWY_EXPORT_AND_TEST_P(FastMathTargetTest, TestFastCos);
 HWY_EXPORT_AND_TEST_P(FastMathTargetTest, TestFastErf);
+HWY_EXPORT_AND_TEST_P(FastMathTargetTest, TestCubeRoot);
 HWY_EXPORT_AND_TEST_P(FastMathTargetTest, TestFastSRGB);
 HWY_EXPORT_AND_TEST_P(FastMathTargetTest, TestFastPQDFE);
 HWY_EXPORT_AND_TEST_P(FastMathTargetTest, TestFastPQEFD);
